requirements.txt
pigpio==1.78
RPi.GPIO==0.7.1
